Rename package "songs" to "tracks"
