+++ app/src/main/java/fr/nihilus/music/library/tracks/AllTracksFragment.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package fr.nihilus.music.library.songs
+package fr.nihilus.music.library.tracks
 
 import android.Manifest
 import android.app.Activity
