+++ app/src/main/java/fr/nihilus/music/library/tracks/TracksViewModel.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package fr.nihilus.music.library.songs
+package fr.nihilus.music.library.tracks
 
 import android.support.v4.media.MediaBrowserCompat
 import androidx.lifecycle.ViewModel
